fix: prevent spurious GET /api/profiles/{slug} on /profile/create page
Extract CreateProfileService from ProfileService so that the create-profile
page no longer triggers ProfileService's profileResource, which would fire
a GET request for a profile that doesn't exist yet (404).
